Split `Context` and `GatewayProtocol` from `Gateway`
