+++ Gateway.php
@@ -14,9 +14,10 @@ use \Exception;
  * @license http://www.opensource.org/licenses/mit-license.php MIT License
  */
 
-/**
- * 数据发送相关
- */
+if (! class_exists(\Composer\Autoload\ClassLoader::class)) {
+    require_once __DIR__ . '/Context.php';
+    require_once __DIR__ . '/GatewayProtocol.php';
+}
 
 /**
  * 数据发送相关
@@ -1372,277 +1373,5 @@ class Gateway
             return false;
         }
         return true;
-    }
-}
-
-
-/**
- * 上下文 包含当前用户uid， 内部通信local_ip local_port socket_id ，以及客户端client_ip client_port
- */
-class Context
-{
-    /**
-     * 内部通讯id
-     * @var string
-     */
-    public static $local_ip;
-    /**
-     * 内部通讯端口
-     * @var int
-     */
-    public static $local_port;
-    /**
-     * 客户端ip
-     * @var string
-     */
-    public static $client_ip;
-    /**
-     * 客户端端口
-     * @var int
-     */
-    public static $client_port;
-    /**
-     * client_id
-     * @var string
-     */
-    public static $client_id;
-    /**
-     * 连接connection->id
-     * @var int
-     */
-    public static $connection_id;
-
-    /**
-     * 旧的session
-     *
-     * @var string
-     */
-    public static $old_session;
-
-    /**
-     * 编码session
-     * @param mixed $session_data
-     * @return string
-     */
-    public static function sessionEncode($session_data = '')
-    {
-        if($session_data !== '')
-        {
-            return serialize($session_data);
-        }
-        return '';
-    }
-
-    /**
-     * 解码session
-     * @param string $session_buffer
-     * @return mixed
-     */
-    public static function sessionDecode($session_buffer)
-    {
-        return unserialize($session_buffer);
-    }
-
-    /**
-     * 清除上下文
-     * @return void
-     */
-    public static function clear()
-    {
-        static::$local_ip = static::$local_port = static::$client_ip = static::$client_port =
-        static::$client_id = static::$connection_id  = static::$old_session = null;
-    }
-
-    /**
-     * 通讯地址到client_id的转换
-     * @return string
-     */
-    public static function addressToClientId($local_ip, $local_port, $connection_id)
-    {
-        return bin2hex(pack('NnN', $local_ip, $local_port, $connection_id));
-    }
-
-    /**
-     * client_id到通讯地址的转换
-     * @return array
-     */
-    public static function clientIdToAddress($client_id)
-    {
-        if(strlen($client_id) !== 20)
-        {
-            throw new \Exception("client_id $client_id is invalid");
-        }
-        return unpack('Nlocal_ip/nlocal_port/Nconnection_id' ,pack('H*', $client_id));
-    }
-
-}
-
-
-/**
- * Gateway 与 Worker 间通讯的二进制协议
- *
- * struct GatewayProtocol
- * {
- *     unsigned int        pack_len,
- *     unsigned char       cmd,//命令字
- *     unsigned int        local_ip,
- *     unsigned short      local_port,
- *     unsigned int        client_ip,
- *     unsigned short      client_port,
- *     unsigned int        connection_id,
- *     unsigned char       flag,
- *     unsigned short      gateway_port,
- *     unsigned int        ext_len,
- *     char[ext_len]       ext_data,
- *     char[pack_length-HEAD_LEN] body//包体
- * }
- * NCNnNnNCnN
- */
-class GatewayProtocol
-{
-    // 发给worker，gateway有一个新的连接
-    const CMD_ON_CONNECT = 1;
-    // 发给worker的，客户端有消息
-    const CMD_ON_MESSAGE = 3;
-    // 发给worker上的关闭链接事件
-    const CMD_ON_CLOSE = 4;
-    // 发给gateway的向单个用户发送数据
-    const CMD_SEND_TO_ONE = 5;
-    // 发给gateway的向所有用户发送数据
-    const CMD_SEND_TO_ALL = 6;
-    // 发给gateway的踢出用户
-    // 1、如果有待发消息，将在发送完后立即销毁用户连接
-    // 2、如果无待发消息，将立即销毁用户连接
-    const CMD_KICK = 7;
-    // 发给gateway的立即销毁用户连接
-    const CMD_DESTROY = 8;
-    // 发给gateway，通知用户session更新
-    const CMD_UPDATE_SESSION = 9;
-    // 获取在线状态
-    const CMD_GET_ALL_CLIENT_SESSIONS = 10;
-    // 判断是否在线
-    const CMD_IS_ONLINE = 11;
-    // client_id绑定到uid
-    const CMD_BIND_UID = 12;
-    // 解绑
-    const CMD_UNBIND_UID = 13;
-    // 向uid发送数据
-    const CMD_SEND_TO_UID = 14;
-    // 根据uid获取绑定的clientid
-    const CMD_GET_CLIENT_ID_BY_UID = 15;
-    // 批量获取uid列表批量获取绑定的clientid
-    const CMD_BATCH_GET_CLIENT_ID_BY_UID = 16;
-    // 加入组
-    const CMD_JOIN_GROUP = 20;
-    // 离开组
-    const CMD_LEAVE_GROUP = 21;
-    // 向组成员发消息
-    const CMD_SEND_TO_GROUP = 22;
-    // 获取组成员
-    const CMD_GET_CLIENT_SESSIONS_BY_GROUP = 23;
-    // 获取组在线连接数
-    const CMD_GET_CLIENT_COUNT_BY_GROUP = 24;
-    // 按照条件查找
-    const CMD_SELECT = 25;
-    // 获取在线的群组ID
-    const CMD_GET_GROUP_ID_LIST = 26;
-    // 取消分组
-    const CMD_UNGROUP = 27;
-    // worker连接gateway事件
-    const CMD_WORKER_CONNECT = 200;
-    // 心跳
-    const CMD_PING = 201;
-    // GatewayClient连接gateway事件
-    const CMD_GATEWAY_CLIENT_CONNECT = 202;
-    // 根据client_id获取session
-    const CMD_GET_SESSION_BY_CLIENT_ID = 203;
-    // 发给gateway，覆盖session
-    const CMD_SET_SESSION = 204;
-    // 当websocket握手时触发，只有websocket协议支持此命令字
-    const CMD_ON_WEBSOCKET_CONNECT = 205;
-    // 包体是标量
-    const FLAG_BODY_IS_SCALAR = 0x01;
-    // 通知gateway在send时不调用协议encode方法，在广播组播时提升性能
-    const FLAG_NOT_CALL_ENCODE = 0x02;
-    /**
-     * 包头长度
-     *
-     * @var int
-     */
-    const HEAD_LEN = 28;
-    public static $empty = array(
-        'cmd'           => 0,
-        'local_ip'      => 0,
-        'local_port'    => 0,
-        'client_ip'     => 0,
-        'client_port'   => 0,
-        'connection_id' => 0,
-        'flag'          => 0,
-        'gateway_port'  => 0,
-        'ext_data'      => '',
-        'body'          => '',
-    );
-    /**
-     * 返回包长度
-     *
-     * @param string $buffer
-     * @return int return current package length
-     */
-    public static function input($buffer)
-    {
-        if (strlen($buffer) < self::HEAD_LEN) {
-            return 0;
-        }
-        $data = unpack("Npack_len", $buffer);
-        return $data['pack_len'];
-    }
-    /**
-     * 获取整个包的 buffer
-     *
-     * @param mixed $data
-     * @return string
-     */
-    public static function encode($data)
-    {
-        $flag = (int)is_scalar($data['body']);
-        if (!$flag) {
-            $data['body'] = serialize($data['body']);
-        }
-        $data['flag'] |= $flag;
-        $ext_len      = strlen($data['ext_data']);
-        $package_len  = self::HEAD_LEN + $ext_len + strlen($data['body']);
-        return pack("NCNnNnNCnN", $package_len,
-                $data['cmd'], $data['local_ip'],
-                $data['local_port'], $data['client_ip'],
-                $data['client_port'], $data['connection_id'],
-                $data['flag'], $data['gateway_port'],
-                $ext_len) . $data['ext_data'] . $data['body'];
-    }
-    /**
-     * 从二进制数据转换为数组
-     *
-     * @param string $buffer
-     * @return array
-     */
-    public static function decode($buffer)
-    {
-        $data = unpack("Npack_len/Ccmd/Nlocal_ip/nlocal_port/Nclient_ip/nclient_port/Nconnection_id/Cflag/ngateway_port/Next_len",
-            $buffer);
-        if ($data['ext_len'] > 0) {
-            $data['ext_data'] = substr($buffer, self::HEAD_LEN, $data['ext_len']);
-            if ($data['flag'] & self::FLAG_BODY_IS_SCALAR) {
-                $data['body'] = substr($buffer, self::HEAD_LEN + $data['ext_len']);
-            } else {
-                $data['body'] = unserialize(substr($buffer, self::HEAD_LEN + $data['ext_len']));
-            }
-        } else {
-            $data['ext_data'] = '';
-            if ($data['flag'] & self::FLAG_BODY_IS_SCALAR) {
-                $data['body'] = substr($buffer, self::HEAD_LEN);
-            } else {
-                $data['body'] = unserialize(substr($buffer, self::HEAD_LEN));
-            }
-        }
-        return $data;
     }
 }
